move user mutation tests
