+++ api-js/src/mutations/user/__tests__/send-email-verification.test.js
@@ -5,14 +5,14 @@ const bcrypt = require('bcrypt')
 const { graphql, GraphQLSchema } = require('graphql')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { createQuerySchema } = require('../queries')
-const { createMutationSchema } = require('../mutations')
-const { cleanseInput } = require('../validators')
-const { tokenize } = require('../auth')
-const { userLoaderByUserName } = require('../loaders')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { createQuerySchema } = require('../../../queries')
+const { createMutationSchema } = require('../..')
+const { cleanseInput } = require('../../../validators')
+const { tokenize } = require('../../../auth')
+const { userLoaderByUserName } = require('../../../loaders')
 
 const mockNotify = jest.fn()
